+++ client/src/views/pages/Store/StorePage.vue
@@ -174,6 +174,8 @@ export default {
     await StoreData.mutations.loadStoreInformation(this.storeId)
     await StoreData.mutations.loadGetRegionOptions(this.storeInformation.region.id)
     await StoreData.mutations.loadStoreMember(this.storeId)
+    await StoreData.mutations.loadStoreLog(this.storeId)
+
     this.checkIsUserInStore()
     this.getLastFetchDate()
     this.loadRightsInfo()